--- src/routes/AddAccount/index.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router'
-import LinkAccountModule from '../../Components/LinkAccountModule'
-
-export const Route = createFileRoute('/AddAccount/')({
-  component: LinkAccountModule
-})
