+++ src/routes/LinkAccount/index.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from '@tanstack/react-router'
+import LinkAccountModule from '../../Components/LinkAccount/LinkAccountModule'
+
+export const Route = createFileRoute('/LinkAccount/')({
+  component: LinkAccountModule
+})
